HomeWork №1
Step 1: Install Oracle Virtual Box Manager, then select "create a new virtual machine". In the drop-down window, I set the name of the virtual machine, its location, and also specified the path to the image of the Windows 10 x64 operating system.

Step 2: In the next window, set the username, password, and specified the product key.

Step 3: I allocated the necessary amount of RAM, in my opinion, as well as the number of virtual processors (previously in UEFI it is necessary to enable the possibility of core virtualization in the CPU settings), and click next.

Step 4: At this stage, I created a virtual hard drive and allocated 50GB of physical memory to my computer.

Step 5: Checked all the installations that I have made up to this point and confirmed by clicking done.

Step 6, 7 and 8: The VM was created successfully. Then I made a standard installation of Windows 10. The virtual machine is now ready to work.